moving format merge to dex.ts
 - preforming other cleanup.
