+++ sim/dex.ts
@@ -47,12 +47,14 @@ import * as path from 'path';
 import * as Data from './dex-data';
 import {PRNG, PRNGSeed} from './prng';
 import {Utils} from '../lib/utils';
+import { FormatList } from '../config/custom-formats';
 
 const BASE_MOD = 'gen8' as ID;
 const DEFAULT_MOD = BASE_MOD;
 const DATA_DIR = path.resolve(__dirname, '../.data-dist');
 const MODS_DIR = path.resolve(__dirname, '../.data-dist/mods');
-const FORMATS = path.resolve(__dirname, '../.config-dist/formats');
+const MAINFORMATS = path.resolve(__dirname, '../.config-dist/formats');
+const CUSTOMFORMATS = path.resolve(__dirname, '../.config-dist/custom-formats')
 
 const dexes: {[mod: string]: ModdedDex} = Object.create(null);
 
@@ -131,6 +133,77 @@ const Natures: {[k: string]: Nature} = {
 	timid: {name: "Timid", plus: 'spe', minus: 'atk'},
 };
 export const toID = Data.toID;
+
+// interface for the builder.
+interface FormatSection {
+	section: string;
+	column?: number;
+	formats: FormatData[];
+}
+
+// function for merging the two lists
+function merge(
+			   main: FormatList[],
+			   side: FormatList[]
+): FormatList[] {
+	// result that is return and makes the actual list for formats.
+	const result: FormatList[] = [];
+
+	// used as a intermediary to build the final list.
+	const build: FormatSection[] = [];
+
+	// used to track location to keep formats under their sections.
+	let loc = -1;
+
+	// populates the original sections and formats easily
+	// there should be no repeat sections at this point.
+	for (const element of main) {
+		if (element.section) {
+			build.push({section: element.section, column: element.column, formats: []});
+			loc++;
+		} else if ((element as FormatData).name) {
+			build[loc].formats.push((element as FormatData));
+		}
+	}
+
+	// merges the second list the hard way. Accounts for repeats.
+	for (const element of side) {
+		// finds the section and makes it if it doesn't exist.
+		if (element.section) {
+			loc = 0;
+			let found = false;
+
+			// finds the loc of the section header (or next loc if it's new)
+			for (const entry of build) {
+				if (entry.section === element.section) {
+					found = true;
+					break;
+				}
+				loc++;
+			}
+
+			// if it's new it makes a new entry.
+			if (!found) {
+				build.push({section: element.section, column: element.column, formats: []});
+			}
+		} else if ((element as FormatData).name) { // otherwise, adds the element to its section.
+			build[loc].formats.push(element as FormatData);
+		}
+	}
+
+	// builds the final result.
+	for (const element of build) {
+		// adds the section to the list.
+		result.push({section: element.section, column: element.column});
+
+		// adds all the formats in the section.
+		for (const entry of element.formats) {
+			result.push(entry);
+		}
+	}
+
+	return result;
+}
 
 export class ModdedDex {
 	readonly Data: typeof Data;
@@ -1518,7 +1591,7 @@ export class ModdedDex {
 		// Load formats
 		let Formats;
 		try {
-			Formats = require(FORMATS).Formats;
+			Formats = (merge(require(MAINFORMATS).MainFormats, require(CUSTOMFORMATS).CustomFormats) as any);
 		} catch (e) {
 			if (e.code !== 'MODULE_NOT_FOUND' && e.code !== 'ENOENT') {
 				throw e;
